Resampling pixels moved to raster service
